+++ Services/AppVersionProvider.cs
@@ -4,8 +4,8 @@ namespace UrbanPlanToolbox.Services;
 
 public static class AppVersionProvider
 {
-    public const string Version = "1.2.0";
-    public const string DisplayVersion = "v1.2.0";
+    public const string Version = "1.2.1";
+    public const string DisplayVersion = "v1.2.1";
     public const int DataSchemaVersion = 1;
 
     public static Version GetCurrentVersion()